Extract style serialisation helpers to separate file
